+++ packages/chains/firo/lib/firoTransaction.ts
@@ -41,14 +41,15 @@ class FiroTransaction extends PaymentTransaction {
    * @returns json representation of the payment transaction
    */
   toJson = (): string => {
-    return JSON.stringify({
+    const obj = {
       network: this.network,
       eventId: this.eventId,
       txBytes: this.getTxHexString(),
       txId: this.txId,
       txType: this.txType,
       inputUtxos: this.inputUtxos,
-    });
+    };
+    return JSON.stringify(obj, Object.keys(obj).sort());
   };
 
   /**