+++ packages/chains/bitcoin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rosen-chains/bitcoin
 
+## 11.0.0
+
+### Major Changes
+
+- Change transaction object toJson to sort object keys before stringifying
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-chains/abstract-chain@17.0.0
+
 ## 10.0.1
 
 ### Patch Changes