+++ packages/chains/ergo/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rosen-chains/ergo
 
+## 15.0.0
+
+### Major Changes
+
+- Change transaction object toJson to sort object keys before stringifying
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-chains/abstract-chain@17.0.0
+
 ## 14.1.0
 
 ### Minor Changes